Some funciton is added.
PCA_calculation is a parametric PCA algorithm. It will be arranged.

wrongNumberException is a function that the entered number is not an indis on array. It checks these condition and if it is not, it warns the user and it want to re-entered value.

valuePrinting : it is completely some printing statements. It used more than 2 place. So, I coded a basic funciton.